requirements.txt
et-xmlfile==1.1.0
kafka-python==2.0.2
numpy==1.24.3
openpyxl==3.1.2
pandas==2.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.10.0
tzdata==2024.1
